pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.7",
    "langsmith>=0.4.27",
    "wikipedia-api>=0.8.1",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
